Bug 1389230 - Create churn utilities
This module contains utility functions that are separate from the logic
of the churn script.
